requirements.txt
streamlit
msal
requests
pandas
altair
azure-data-tables
azure-storage-blob
PyPDF2
python-docx
python-pptx
openpyxl
chardet
